+++ include/tropicssl/dhm.h
@@ -45,7 +45,7 @@
 #define TROPICSSL_ERR_DHM_CALC_SECRET_FAILED                -0x04D0
 
 typedef struct {
-	int len;		/*!<  size(P) in chars  */
+	size_t len;		/*!<  size(P) in chars  */
 	mpi P;			/*!<  prime modulus     */
 	mpi G;			/*!<  generator         */
 	mpi X;			/*!<  secret value      */
@@ -88,7 +88,7 @@ extern "C" {
 	 * \return         0 if successful, or an TROPICSSL_ERR_DHM_XXX error code
 	 */
 	int dhm_make_params(dhm_context * ctx, int x_size,
-			    unsigned char *output, int *olen,
+			    unsigned char *output, size_t *olen,
 			    int (*f_rng) (void *), void *p_rng);
 
 	/**
@@ -100,7 +100,7 @@ extern "C" {
 	 *
 	 * \return         0 if successful, or an TROPICSSL_ERR_DHM_XXX error code
 	 */
-	int dhm_read_public(dhm_context * ctx, const unsigned char *input, int ilen);
+	int dhm_read_public(dhm_context * ctx, const unsigned char *input, size_t ilen);
 
 	/**
 	 * \brief          Create own private value X and export G^X
@@ -115,7 +115,7 @@ extern "C" {
 	 * \return         0 if successful, or an TROPICSSL_ERR_DHM_XXX error code
 	 */
 	int dhm_make_public(dhm_context * ctx, int s_size,
-			    unsigned char *output, int olen,
+			    unsigned char *output, size_t olen,
 			    int (*f_rng) (void *), void *p_rng);
 
 	/**
@@ -128,7 +128,7 @@ extern "C" {
 	 * \return         0 if successful, or an TROPICSSL_ERR_DHM_XXX error code
 	 */
 	int dhm_calc_secret(dhm_context * ctx,
-			    unsigned char *output, int *olen);
+			    unsigned char *output, size_t *olen);
 
 	/*
 	 * \brief          Free the components of a DHM key
